% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{qscoreOpt}
\alias{qscoreOpt}
\title{qscoreOpt}
\usage{
qscoreOpt(spe = spe, plot = FALSE, custom = FALSE)
}
\arguments{
\item{spe}{A `SpatialExperiment` object with spatial transcriptomics data.}

\item{plot}{a boolean to set to TRUE if you want to see the glm plot}

\item{custom}{a boolean value to set to TRUE if you are using custom polygons}
}
\value{
The `SpatialExperiment` object with added filter flags in `colData`.
}
\description{
Compute automatically weights for quality score through glm training

This function optimizes quality score by computing a glm to determine weights
for the quality score formula after selecting a training set from the
`SpatialExperiment` object.
}
\examples{
# TBD
}
